+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Security
 - Update simpleidn to 0.2.3. CNJR-9498
+- Update rack to 2.2.14 to address CVE-2025-46727. CONJSE-2001
 
 ## [1.22.0] - 2025-04-02