+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Nothing should go in this section, please add to the latest unreleased version
   (and update the corresponding date), or add a new version.
 
+## [1.21.4] - 2025-01-13
+### Security
+- Update rack to 2.2.13 to address CVE-2025-27610.
+  CONJSE-1956
+
 ## [1.21.3] - 2024-12-16
 
 ### Fixed